feat(packagejson): updated version of herbsjs and glues
Updated herbs version to ^2.0.0
Updated herbarium and herbsshelf
Updated herbs2mongo, herbs2knex,
herbs2gql and herbs2rest

BREAKING CHANGE: A feature was implemented in version 2.0.0 of herbs that places the content passed
to the authorize function inside ctx.user.
However in the project generated with herbs-cli the
ctx.user is used in the generated usecases.
Refer to:
https://github.com/herbsjs/buchu/commit/570a4f5ae22e24ae757edd5b2269e3859d0c6af5